+++ lib/phx_bb_web/controllers/old_user_registration_controller.ex
@@ -1,4 +1,4 @@
-defmodule PhxBbWeb.UserRegistrationController do
+defmodule PhxBbWeb.OldUserRegistrationController do
   use PhxBbWeb, :controller
 
   alias PhxBb.Accounts
@@ -19,7 +19,7 @@ defmodule PhxBbWeb.UserRegistrationController do
 
     case Accounts.register_user(user_params) do
       {:ok, user} ->
-        {:ok, _} =
+        {:ok, _} = 
           Accounts.deliver_user_confirmation_instructions(
             user,
             &Routes.user_confirmation_url(conn, :confirm, &1)