+++ lib/phx_bb_web/live/components/user_menu_component.ex
@@ -1,0 +1,4 @@
+defmodule PhxBbWeb.UserMenuComponent do
+  use PhxBbWeb, :live_component
+
+end