+++ lib/phx_bb_web/router.ex
@@ -49,8 +49,9 @@ defmodule PhxBbWeb.Router do
   scope "/", PhxBbWeb do
     pipe_through [:browser, :redirect_if_user_is_authenticated]
 
-    get "/users/register", UserRegistrationController, :new
-    post "/users/register", UserRegistrationController, :create
+    live "/users/register", UserRegistrationLive.New, :new, as: :user_registration
+    # get "/users/register", UserRegistrationController, :new
+    # post "/users/register", UserRegistrationController, :create
     get "/users/log_in", UserSessionController, :new
     post "/users/log_in", UserSessionController, :create
     get "/users/reset_password", UserResetPasswordController, :new